--- a/doc/doxygen/install/install-macs.doxygen
+++ b/doc/doxygen/install/install-macs.doxygen
@@ -128,7 +128,7 @@
   required by %OpenMS.
 
     \code
-      brew install libsvm xerces-c boost glpk eigen sqlite hdf5
+      brew install libsvm xerces-c boost glpk eigen hdf5
     \endcode
 
   @note Until the CoinMP complete package is migrated to the coin-or-tools tap, we suggest to use our contrib version. You will then additionally need gfortran as a Fortran compiler to build.
@@ -139,7 +139,7 @@
   libraries required by %OpenMS.
 
     \code
-      sudo port install libsvm xercesc3 boost glpk eigen3 sqlite hdf5
+      sudo port install libsvm xercesc3 boost glpk eigen3 hdf5
     \endcode
 
  The rest needs to be installed via the contrib.