--- a/src/openms/include/OpenMS/FORMAT/OMSFileLoad.h
+++ b/src/openms/include/OpenMS/FORMAT/OMSFileLoad.h
@@ -45,8 +45,6 @@
 
 #include <QtCore/QJsonArray> // for JSON export
 
-class SQLite::Statement;
-
 namespace OpenMS
 {
   namespace Internal