--- a/src/openms/source/FORMAT/OMSFile.cpp
+++ b/src/openms/source/FORMAT/OMSFile.cpp
@@ -58,13 +58,6 @@ namespace OpenMS
 
   void OMSFile::load(const String& filename, IdentificationData& id_data)
   {
-    /*if (!QCoreApplication::instance())
-    {
-      char argv[1] = "";
-      int argc = 0;
-      QCoreApplication a(argc, reinterpret_cast<char**>(&argv));
-      std::cout << QCoreApplication::libraryPaths().join(" ").toStdString() << std::endl;
-    }*/
     OpenMS::Internal::OMSFileLoad helper(filename, log_type_);
     helper.load(id_data);
   }