@timosachsenberg 2022-11-17
@enetz can you run this through your benchmarking code you did when simplifying the schema?