@timosachsenberg 2022-11-18
Is there a way to just fetch that code instead of including it?
I have no clue but some packages do something like
```
include(FetchContent)
FetchContent_Declare(
  xxx
  GIT_REPOSITORY xxxx
  GIT_TAG        vxxxx
)
FetchContent_MakeAvailable(xxxx)
```