@cbielow 2022-11-22
> Is there a way to just fetch that code instead of including it? I have no clue but some packages do something like
> 
> ```
> include(FetchContent)
> FetchContent_Declare(
>   xxx
>   GIT_REPOSITORY xxxx
>   GIT_TAG        vxxxx
> )
> FetchContent_MakeAvailable(xxxx)
> ```

Never tried actually.
If we move in this direction it would make sense to be consistent here IMHO, and do that with the other thirdparty libs as well.
But this is out of scope of this PR.