@cbielow 2022-11-22
> Do we need to include the example code in third_parties?

Does not hurt (apart from a few KB), and may be a nice reference for lookup in later sqlite implementation efforts.