@cbielow 2022-11-22
ready for merge. Tests pass (even though Windows CI seems down).
Any more comments/TODO's?