@cbielow 2022-11-22
> so libsqlite is now also included through thirdparty (not the contrib anymore)?

Yes. So we could remove it from contrib as well.

Other things which could be done (other PR):
 - remove our own sqliteadapter and replace with SQLiteCPP (mostly affects openswath code)