@timosachsenberg 2022-11-22
k I think this is good to go. I can remove slite from contrib as well as I got a PR open anyways.