@SRombauts 2022-11-23
Thanks @cbielow from pinging me, I am delighted to see your work has been merged, and as such SQLiteCpp